6/Services/Auth/requirements.txt
Flask
pyyaml
pyjwt
flask-selfdoc
